+++ extras/tests/JsonDocument/ElementProxy.cpp
@@ -246,10 +246,10 @@ TEST_CASE("ElementProxy cast to JsonVariant") {
   CHECK(doc.as<std::string>() == "[\"toto\"]");
 }
 
-TEST_CASE("ElementProxy::link()") {
+TEST_CASE("ElementProxy::shallowCopy()") {
   StaticJsonDocument<1024> doc1, doc2;
-  doc1[0].link(doc2);
   doc2["hello"] = "world";
+  doc1[0].shallowCopy(doc2);
 
   CHECK(doc1.as<std::string>() == "[{\"hello\":\"world\"}]");
 }