+++ extras/tests/JsonVariant/nesting.cpp
@@ -28,12 +28,4 @@ TEST_CASE("JsonVariant::nesting()") {
     var.to<JsonArray>();
     REQUIRE(var.nesting() == 1);
   }
-
-  SECTION("returns depth of linked array") {
-    StaticJsonDocument<128> doc2;
-    doc2[0][0] = 42;
-    var.link(doc2);
-
-    CHECK(var.nesting() == 2);
-  }
 }