+++ src/ArduinoJson/Json/JsonSerializer.hpp
@@ -24,7 +24,7 @@ class JsonSerializer : public Visitor<size_t> {
     const VariantSlot *slot = array.head();
 
     while (slot != 0) {
-      slot->data()->resolve()->accept(*this);
+      slot->data()->accept(*this);
 
       slot = slot->next();
       if (slot == 0)
@@ -45,7 +45,7 @@ class JsonSerializer : public Visitor<size_t> {
     while (slot != 0) {
       _formatter.writeString(slot->key());
       write(':');
-      slot->data()->resolve()->accept(*this);
+      slot->data()->accept(*this);
 
       slot = slot->next();
       if (slot == 0)