+++ src/ArduinoJson/Variant/VariantRef.hpp
@@ -117,8 +117,7 @@ class VariantConstRef : public VariantRefBase<const VariantData>,
   }
 
   FORCE_INLINE VariantConstRef getElementConst(size_t index) const {
-    return VariantConstRef(_data != 0 ? _data->resolve()->getElement(index)
-                                      : 0);
+    return VariantConstRef(_data != 0 ? _data->getElement(index) : 0);
   }
 
   FORCE_INLINE VariantConstRef operator[](size_t index) const {
@@ -129,8 +128,7 @@ class VariantConstRef : public VariantRefBase<const VariantData>,
   // getMemberConst(const String&) const
   template <typename TString>
   FORCE_INLINE VariantConstRef getMemberConst(const TString &key) const {
-    return VariantConstRef(_data ? _data->resolve()->getMember(adaptString(key))
-                                 : 0);
+    return VariantConstRef(_data ? _data->getMember(adaptString(key)) : 0);
   }
 
   // getMemberConst(char*) const
@@ -138,8 +136,7 @@ class VariantConstRef : public VariantRefBase<const VariantData>,
   // getMemberConst(const __FlashStringHelper*) const
   template <typename TChar>
   FORCE_INLINE VariantConstRef getMemberConst(TChar *key) const {
-    return VariantConstRef(_data ? _data->resolve()->getMember(adaptString(key))
-                                 : 0);
+    return VariantConstRef(_data ? _data->getMember(adaptString(key)) : 0);
   }
 
   // operator[](const std::string&) const
@@ -281,8 +278,7 @@ class VariantRef : public VariantRefBase<VariantData>,
   }
 
   FORCE_INLINE VariantConstRef getElementConst(size_t index) const {
-    return VariantConstRef(_data != 0 ? _data->resolve()->getElement(index)
-                                      : 0);
+    return VariantConstRef(_data != 0 ? _data->getElement(index) : 0);
   }
 
   FORCE_INLINE VariantRef getOrAddElement(size_t index) const {
@@ -310,8 +306,7 @@ class VariantRef : public VariantRefBase<VariantData>,
   // getMemberConst(const __FlashStringHelper*) const
   template <typename TChar>
   FORCE_INLINE VariantConstRef getMemberConst(TChar *key) const {
-    return VariantConstRef(_data ? _data->resolve()->getMember(adaptString(key))
-                                 : 0);
+    return VariantConstRef(_data ? _data->getMember(adaptString(key)) : 0);
   }
 
   // getMemberConst(const std::string&) const
@@ -320,8 +315,7 @@ class VariantRef : public VariantRefBase<VariantData>,
   FORCE_INLINE
       typename enable_if<IsString<TString>::value, VariantConstRef>::type
       getMemberConst(const TString &key) const {
-    return VariantConstRef(_data ? _data->resolve()->getMember(adaptString(key))
-                                 : 0);
+    return VariantConstRef(_data ? _data->getMember(adaptString(key)) : 0);
   }
 
   // getOrAddMember(char*) const
@@ -361,12 +355,12 @@ class VariantRef : public VariantRefBase<VariantData>,
       _data->remove(adaptString(key));
   }
 
-  inline void link(VariantConstRef target) {
+  inline void shallowCopy(VariantConstRef target) {
     if (!_data)
       return;
     const VariantData *targetData = getData(target);
     if (targetData)
-      _data->setPointer(targetData);
+      *_data = *targetData;
     else
       _data->setNull();
   }