+++ CHANGELOG.md
@@ -4,7 +4,7 @@ ArduinoJson: change log
 HEAD
 ----
 
-* Add `JsonVariant::link()` (issue #1343)
+* Add `JsonVariant::shallowCopy()` (issue #1343)
 * Fix `9.22337e+18 is outside the range of representable values of type 'long'`
 * Fix comparison operators for `JsonArray`, `JsonArrayConst`, `JsonObject`, and `JsonObjectConst`
 * Remove undocumented `accept()` functions